backend/requirements.txt
open-interpreter>=0.3.0
fastapi>=0.110,<1.0
uvicorn[standard]>=0.29,<1.0
websockets>=12.0,<13.0
python-dotenv>=1.0,<2.0
python-telegram-bot>=21.0,<22.0
anthropic>=0.25,<1.0
PyYAML>=6.0,<7.0
aiofiles>=23.0,<24.0
